+++ src/main/java/ru/itmo/javaadvanced/lesson1/ApplicationRunner.java
@@ -1,11 +1,11 @@
-package ru.itmo.javaadvanced;
+package ru.itmo.javaadvanced.lesson1;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import ru.itmo.javaadvanced.dto.ExampleDto;
+import ru.itmo.javaadvanced.lesson1.dto.ExampleDto;
 
 @SpringBootApplication
-public class JavaadvancedApplication {
+public class ApplicationRunner {
 
 	private static final ExampleDto EXAMPLE_DTO = new ExampleDto();
 
@@ -13,7 +13,7 @@ public class JavaadvancedApplication {
 		EXAMPLE_DTO.setId(1L);
 		System.out.println(EXAMPLE_DTO.getId());
 
-		SpringApplication.run(JavaadvancedApplication.class, args);
+		SpringApplication.run(ApplicationRunner.class, args);
 	}
 
 }
