+++ src/main/java/ru/itmo/javaadvanced/lesson1/dto/ExampleDto.java
@@ -1,4 +1,4 @@
-package ru.itmo.javaadvanced.dto;
+package ru.itmo.javaadvanced.lesson1.dto;
 
 import lombok.Data;
