+++ src/test/java/ru/itmo/javaadvanced/lesson1/ApplicationRunnerTests.java
@@ -1,10 +1,10 @@
-package ru.itmo.javaadvanced;
+package ru.itmo.javaadvanced.lesson1;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class JavaadvancedApplicationTests {
+class ApplicationRunnerTests {
 
 	@Test
 	void contextLoads() {
